V5.4-000

+++ sr_port/getstorage.h
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001, 2007 Fidelity Information Services, Inc	*
+ *	Copyright 2001, 2010 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -12,6 +12,6 @@
 #ifndef GETSTORAGE_INCLUDED
 #define GETSTORAGE_INCLUDED
 
-ssize_t getstorage(void);
+int4 getstorage(void);
 
 #endif /* GETSTORAGE_INCLUDED */
